avs_commons 4.9.0
Features:
- Support for completely replacing the avs_log implementation
- Better support for integrating with custom TLS compatibility layers
- (commercial version only) Support for PSA API for hardware-based security

Improvements:
- Added support for Mbed TLS 3.0

Bugfixes:
- Fixed compatibility some platforms, including ESP-IDF and MinGW (winpthreads)

+++ src/persistence/avs_persistence.c
@@ -129,7 +129,7 @@ static avs_error_t persist_sized_buffer(avs_persistence_context_t *ctx,
         LOG(ERROR,
             _("Element too big to persist (") "%lu" _(
                     " is larger than ") "%" PRIu32 _(")"),
-            (unsigned long) *size_ptr, UINT32_MAX);
+            (unsigned long) *size_ptr, (uint32_t) UINT32_MAX);
         return avs_errno(AVS_EOVERFLOW);
     }
     avs_error_t err = persist_u32(ctx, &size32);
